Made it work in the new update and added support for jpg files.
very silly
