--- Source Code/PluginInfo.cs
@@ -1,12 +1,0 @@
-﻿namespace Computerbackgrounds
-{
-    /// <summary>
-    /// This class is used to provide information about your mod to BepInEx.
-    /// </summary>
-    internal class info
-    {
-        public const string GUID = "com.strikerangelvibes.gorillatag.computerbackgrounds";
-        public const string Name = "Computerbackgrounds";
-        public const string Version = "1.0.0";
-    }
-}
