+++ Source Code/computer-backgrounds/PluginInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace Computerbackgrounds
+{
+    internal class info
+    {
+        public const string GUID = "com.strikerangelvibes.gorillatag.computerbackgrounds";
+        public const string Name = "Computerbackgrounds";
+        public const string Version = "1.0.0";
+    }
+}
